[Feat] #51 공지사항 및 분실물 수정(이미지, 컨텐츠)

+++ src/main/java/gdg/festa/application/dto/reserve/ReadReserveStateDto.java
@@ -5,7 +5,7 @@ import lombok.Builder;
 
 @Builder
 public record ReadReserveStateDto(
-        Long waitTeam, // 앞의 대기 팀
+        Integer waitTeam, // 앞의 대기 팀
         ReserveStatus reserveStatus // 사용자의 예약 상태 (ex. 예약중, 예약 취소)
 ) {
 }
